Fix index comparisons to  first row & column
This update will change the row and column index comparisons in order to
check for open spots (or filled spots) in row 0 (when a current
position's y index equals 1) and column 0 (when a current position's x
index equals 1). The previous index comparisons were technically valid,
but they were overly cautious in avoiding IndexOutOfRange exceptions.

See the code diff for details.